requirements.txt
alembic==1.3.1
astroid==2.4.1
Babel==2.8.0
backcall==0.1.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
decorator==4.4.1
dnspython==1.16.0
dominate==2.4.0
email-validator==1.1.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.10.0
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
future==0.18.2
guess-language-spirit==0.5.3
idna==2.8
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.7.0
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.2
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.5.2
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-jsonrpc-server==0.3.2
python-language-server==0.31.2
pytz==2019.3
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.11
style==1.1.0
toml==0.10.0
traitlets==4.3.3
ujson==1.35
update==0.0.1
urllib3==1.25.8
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.12.1
WTForms==2.2.1
